requirements.txt
langchain
langchain-community
oci
psycopg2-binary
pgvector
streamlit
python-dotenv
faker
numpy